// Aplicaciones WEB II/segundo parcial/complementarios/complementario06/vite-project/src/interfaces/ICliente.ts
// Generated by https://quicktype.io

export interface IResCliente {
    total:    number;
    clientes: Cliente[];
}

export interface Cliente {
    _id?:        string;
    nombre:     string;
    apellido:   string;
    estado?:     boolean;
    telefono:   string;
    direccion: string;
    cedula:     string;
    usuario?:    string;
    password?:   string;
    __v?:        number;
}